Add RV64 support to BGBCC.
Added the ability to compile binaries for RV64G mode in BGBCC, partly as I may need this for other things I want to mess with... Still pretty buggy / experimental...

+++ tk_qsrc/doomsrc2/r_state.h
@@ -114,7 +114,7 @@ extern player_t*	viewplayer;
 // ?
 extern angle_t		clipangle;
 
-extern int		viewangletox[FINEANGLES/2];
+extern short		viewangletox[FINEANGLES/2];
 extern angle_t		xtoviewangle[SCREENWIDTH+1];
 //extern fixed_t		finetangent[FINEANGLES/2];
